+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: nthimoni <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 17:31:23 by nthimoni          #+#    #+#             */
-/*   Updated: 2021/11/23 17:31:28 by nthimoni         ###   ########.fr       */
+/*   Updated: 2021/11/24 22:03:35 by nthimoni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memset(void *pointer, int value, size_t count)
 {
-	unsigned int	i;
-	char 		*ptr;
-	
+	size_t	i;
+	char	*ptr;
+
 	if (pointer == NULL)
 		return (NULL);
 	ptr = (char *)pointer;
