+++ ft_split.c
@@ -6,7 +6,7 @@
 /*   By: nthimoni <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/24 16:19:56 by nthimoni          #+#    #+#             */
-/*   Updated: 2021/11/24 17:09:17 by nthimoni         ###   ########.fr       */
+/*   Updated: 2021/11/24 22:03:09 by nthimoni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 static	int	wrdcnt(char const *s, char c)
 {
-	int count;
-	int i;
+	int	count;
+	int	i;
 
 	i = 0;
 	count = 0;
-	while(s[i])
+	while (s[i])
 	{
 		while (s[i] && s[i] == c)
 			i++;
@@ -45,7 +45,7 @@ static	char	*cpywrd(char const *s, char c)
 {
 	char	*wrd;
 	size_t	len;
-	
+
 	len = wrdlen(s, c);
 	wrd = malloc(sizeof(char) * (len + 1));
 	ft_memcpy(wrd, s, len);
@@ -53,12 +53,12 @@ static	char	*cpywrd(char const *s, char c)
 	return (wrd);
 }
 
-char		**ft_split(char const *s, char c)
+char	**ft_split(char const *s, char c)
 {
-	char		**ret;
-	unsigned int	word_count;
-	unsigned int	i;
-	unsigned int	a;
+	char	**ret;
+	size_t	word_count;
+	size_t	i;
+	size_t	a;
 
 	word_count = wrdcnt(s, c);
 	ret = malloc(sizeof(char *) * (word_count + 1));
@@ -76,16 +76,3 @@ char		**ft_split(char const *s, char c)
 	ret[word_count] = NULL;
 	return (ret);
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
